+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.0.2](https://gitlab.com/AquaticNightmare/bundle_exporter/-/releases/2.0.1) - 2020-06-13
+## [2.1.0](https://gitlab.com/AquaticNightmare/bundle_exporter/-/releases/2.1.0) - 2020-06-13
+### Added
+- New modifiers:
+    - "Export actions as files" (only FBX) creates separate fbx files with an armature and an animation each (for game development)
+- "Merge Armatures" modifier now deletes the created actions after exporting
+
 ### Fixed
 - Drivers from sub collections are now kept when using "Instanced collections to objects" modifier
+- Fixed error when trying to select a bundle with one of its objects hidden by an excluded collection
 
 ## [2.0.1](https://gitlab.com/AquaticNightmare/bundle_exporter/-/releases/2.0.1) - 2020-05-30
 ### Added